Fix internal method links
Also includes major refactoring of project files.

[closes #554]
