+++ lib/gcloud/jsondoc/version.rb
@@ -1,5 +1,5 @@
 module Gcloud
-  class Jsondoc
+  module Jsondoc
     VERSION = "0.1.0"
   end
 end
